Use uint32_t integer type
uint32_t is a numeric type that guarantees 32 bits.

+++ src/db.cpp
@@ -8,6 +8,7 @@
 #include <boost/version.hpp>
 #include <boost/filesystem.hpp>
 #include <boost/filesystem/fstream.hpp>
+#include <stdint.h>
 
 #include "util.h"
 #include "main.h"
@@ -41,7 +42,7 @@ void CDBEnv::EnvShutdown()
     if (ret != 0)
         printf("EnvShutdown exception: %s (%d)\n", DbEnv::strerror(ret), ret);
     if(!fMockDb)
-      DbEnv(0U).remove(strPath.c_str(), 0);
+      DbEnv((uint32_t)0).remove(strPath.c_str(), 0);
 }
 
 CDBEnv::CDBEnv() : dbenv(DB_CXX_NO_EXCEPTIONS)
